Add support for read-only mode
In read-only mode, any query that is made through .request() will be
blocked if it is not a GET or HEAD request.  A user of the client would
still be able to use the .raw_request() method, but doing so would
typically require that they manage headers themselves.  One place where
this might not have the expected behavior is the uploading of
attachments.  That upload goes through .raw_request() to avoid having
the authorization header sent to the storage server.