Add support for saving Cell changes
I'm not really thrilled with the way this turned out.
But, the containerish nature or Rows and Cells seems like it's worth
keeping even so.

The changes are pretty wide-sweeping.  This is in no small part due
to the fact that when I started writing this library, making the Cells
editable wasn't really a priority.  At all.

But, assigning values to Cells works in at least my limitted testing.