Basic Sheet creation works
I had to rewrite the Column class since instances of it are no longer
created as a result of calls to the API.

There's a whole bunch of work that is left to be done on Sheet
creation, but the base functionality works.
However, the created Sheet has no Rows, and the SDK doesn't yet let you
add Rows to a Sheet.  Baby steps.

Sheet creation looks like:

sc.createSheet("New Sheet 3", [
    smartsheetclient.Column('Task', primary=True),
    smartsheetclient.Column('Start', type=smartsheetclient.CellTypes.Date),
    smartsheetclient.Column('Difficulty', type=smartsheetclient.CellTypes.Picklist)
    ])