Make Row addition work, and other cleanups
Row addition now mostly works.  Where it falls down is when the added
Row changes the numbering of other Rows in the Sheet -- the local Sheet
instance ends up out of sync.  For example, if you do:
        a_sheet.addRow(a_row, position='toTop')
Afer that returns, the local Sheet will have two Rows with rowNumber=1.
That is almost certainly not what you want.
This can be dealt with by refetching the Sheet after each rowAdd().
That's not an ideal solution by any stretch.

I have also added a second mechanism by which Rows and Cells have their
data aggregated for delivery to the API server.  Right now, that
mechanism is used for Row insertion and the prior mechanism is used when
saving changes to a pre-existing Row or Cell.  I need to determine if I
can settle on a single mechanism.

The SheetRows class has been dropped.  I moved its functioinality into
the Sheet class and that makes things a little bit cleaner overall.