Add some real tests and some placeholders
Finally, we have some test cases.  You need to supply each of them
with the path to a file containing the API token.  I don't have anything
in setup.py to automate running the tests either.  That needs to happen
too.

The placeholder tests will be fleshed out.  They serve as a reminder of
what needs to be tested.