Add initial support for links
Basic, initial support for CellHyperlink and CellLinkIn objects.
The Cell object can now properly handle these when received from the
API.

However, it still doesn't deal with the fact that a CellLinkIn can't be
set in a Cell until the Row that Cell is on exists.  I'm not sure, yet,
whether the library should hide that (by saving the Row twice under the
hood), or whether it should expose this constraint with more grace.