Major changes dealing with caching
This is a rather huge set of changes.
At it sheart is an attempt to produce something like a rational approach
to working with cached data when making changes to structural data
(such as adding or deleting Rows or Columns).

I *think* that the interface presented by Sheet, Row, and Column are now
both usable and correct -- and less likely to get the user into rough
patches by making available potentially incorrect cached data.

About the only thing that has changed that is tricky, is the need to
specify that a Sheet should have caching enabled if you want to have new
Cell values available for use prior to saving the changed Row.

I've also done a better job of abstracting out the queries to the API
server.  Since I was no longer working with a "cache all the things!"
model, it sorta had to be done lest the code get even worse.