Reduced # of unnecessary cache misses
I had been hitting the API server all the time to get the columns
information.  This wasn't necessary.  Now, if it makes sense to cache it
off, we cache it off.  In particular, a RowWrapper will get the current
columns information and use that for all of the Rows that it makes.

I added a significant amount of debugging centered around tracking down
the calls to get the columns information.  They can be removed once I am
confident that the logic around when to fetch the columns information
and how to do so is correct.  I *think* it is, but my test cases are no
where near broad and deep enough to be convinced of it just yet.