Beginning work on splitting up client
The client library had gotten so big and unweildy that it was a mess to
try to do anything with it.

So, I've been splitting it up.
Along the way, I have been addressing a few other issues (even though I
know that I shouldn't be mingling work).

Cells are now immutable.  Saving them doesn't work yet, since the Row
change hasn't been resolved, but their code is significantly cleaner.
In particular, the mess that was Cell assignment is now almost nice.