Make initialization work
When creating a Sheet, you need to create the Columns first.  Therefore,
the thinko that had the Column __init__() method taking a required Sheet
instance had to be fixed.

Modified the serialization so that the Column index is only sent if it
is greater than 0 (locally created Columns have their index set to -1
which the server doesn't want to see).