Make Cell.assign() return the new Cell if it can
Cells are immutable, that's going to take some getting used to.
However, Cell.assign() now returns the new Cell -- if it can.

When can't it? when the immediate flag is used to save the Cell
immediately after assigning a value to it.  When that happens, the
entire Row cache gets invalidated, so we really have nothing to return
-- unless we add some complex machinery in Cell.assign().

Frankly, the real solution is probably to eliminate immediate as a
parameter to Cell.assign().  It just makes things more complicated.