Cell assignment no longer saves & strictness
Assigning to a Cell now does not provide the option of immediately
saving the new Cell.  This makes working with Cells more consistent.

Along the way, I've modified the notion of strictness -- it is really an
attribute of each Cell more than it is an attribute of the act of
saving.  As such, it's now part of the Cell, and something that you can
override when saving the Cell (or Row, since Cells are saved by saving
the Row they are on).