Use an email address that's not valid
Ted found an interesting corner case with these test cases.  For
ContactList Cells, if the email address entered is "known" to account, then
the .value of the Cell will be the name corresponding to the email
address.  Otherwise, the .value will be the email address.

Access and testing via different API access tokens makes different
contact list information available, so what "worked on my machine"
didn't work on his.

I've switched to an email address that *shouldn't* ever be marked as
valid -- the "example.com" domain is not a legal domain name.