modified Row.save() to accept a list of cells
The real purpose of this change is to allow saving row properties without saving any Cells, but it might allow saving multiple Cells.  I was running into the problem that I couldn't update (the "collapsed") property on
 my rows because the rows contained formulae, system column types, links, etc.  By passing in an empty list t
o Row.save(cell=[]), I was able to save my row without trying to save any Cells (some of which have values which cannot be saved via the API).