Make Row.save() work with system columns
Eric Chen submitted pull request 27 that in commit
96d94d11568f22146eeecbcfe39b9e7bc4ec2a08 worked around a bug in the way
Row.saveUdate() selected the Cells to send to the server.

This commit fixes the underlying bug in Row.saveUpdate(), and I *think*
addresses the issues from that commit more inline with my original
intent for `Row.save()` and `Row.saveUpdate()`.  That's not to say that
my original intent is definitely correct -- just that this keeps
`Row.save()` conforming to my notion that it's used to save the Row and
potentially add a single Cell to it.

At some point, it would be good to have a discussion about the logic in
`Row.save()` an its iteraction with `Cell.save()`.  I'm still not
convinced that I really have the interface "good" here.  It might be
"good enough", but I'm not comfortable with calling it good.